move service to treatments module
