+++ src/interface/src/app/treatments/treatments.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { TreatmentsRoutingModule } from './treatments-routing.module';
 import { ReviewTreatmentPlanDialogComponent } from './review-treatment-plan-dialog/review-treatment-plan-dialog.component';
+import { TreatmentToPDFService } from 'src/app/treatments/treatment-to-pdf.service';
 
 @NgModule({
   imports: [
@@ -9,5 +10,6 @@ import { ReviewTreatmentPlanDialogComponent } from './review-treatment-plan-dial
     TreatmentsRoutingModule,
     ReviewTreatmentPlanDialogComponent,
   ],
+  providers: [TreatmentToPDFService],
 })
 export class TreatmentsModule {}
